Separate simpl fold and layer ordering enumeration.
