+++ src/main/java/oripa/domain/fold/Folder.java
@@ -29,8 +29,6 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import javax.vecmath.Vector2d;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -45,7 +43,6 @@ import oripa.domain.fold.stackcond.StackConditionOf4Faces;
 import oripa.domain.fold.subface.SubFace;
 import oripa.domain.fold.subface.SubFacesFactory;
 import oripa.geom.GeomUtil;
-import oripa.geom.Line;
 import oripa.util.Matrices;
 import oripa.util.StopWatch;
 import oripa.value.OriLine;
@@ -71,12 +68,15 @@ public class Folder {
 	// helper object
 	private final FolderTool folderTool = new FolderTool();
 
+	private final SimpleFolder simpleFolder;
+
 	private int callCount;
 	private int penetrationTestCallCount;
 	private int penetrationCount;
 
-	public Folder(final SubFacesFactory subFacesFactory) {
+	public Folder(final SubFacesFactory subFacesFactory, final SimpleFolder simpleFolder) {
 		this.subFacesFactory = subFacesFactory;
+		this.simpleFolder = simpleFolder;
 	}
 
 	/**
@@ -99,7 +99,7 @@ public class Folder {
 
 		var foldedModel = new FoldedModel(origamiModel, overlapRelationList);
 
-		simpleFoldWithoutZorder(faces, edges);
+		simpleFolder.simpleFoldWithoutZorder(origamiModel);
 		folderTool.setFacesOutline(faces);
 		List<OriFace> sortedFaces = origamiModel.getSortedFaces();
 		sortedFaces.addAll(faces);
@@ -912,118 +912,6 @@ public class Folder {
 		return bChanged;
 	}
 
-	private void simpleFoldWithoutZorder(
-			final List<OriFace> faces, final List<OriEdge> edges) {
-
-		int id = 0;
-		for (OriFace face : faces) {
-//			face.faceFront = true;
-//			face.movedByFold = false;
-			face.setFaceID(id);
-			id++;
-		}
-
-		walkFace(faces.get(0));
-
-		for (OriEdge e : edges) {
-			var sv = e.getStartVertex();
-			var ev = e.getEndVertex();
-
-			sv.getPosition().set(e.getLeft().getPositionWhileFolding());
-
-			var right = e.getRight();
-			if (right != null) {
-				ev.getPosition().set(right.getPositionWhileFolding());
-			}
-		}
-	}
-
-	// Recursive method that flips the faces, making the folds
-	private void walkFace(final OriFace face) {
-		face.setMovedByFold(true);
-
-		face.halfedgeStream().forEach(he -> {
-			var pair = he.getPair();
-			if (pair == null) {
-				return;
-			}
-			var pairFace = pair.getFace();
-			if (pairFace.isMovedByFold()) {
-				return;
-			}
-
-			flipFace(pairFace, he);
-			pairFace.setMovedByFold(true);
-			walkFace(pairFace);
-		});
-	}
-
-	private void transformVertex(final Vector2d vertex, final Line preLine,
-			final Vector2d afterOrigin, final Vector2d afterDir) {
-		double param[] = new double[1];
-		double d0 = GeomUtil.distance(vertex, preLine, param);
-		double d1 = param[0];
-
-		Vector2d footV = new Vector2d(afterOrigin);
-		footV.x += d1 * afterDir.x;
-		footV.y += d1 * afterDir.y;
-
-		Vector2d afterDirFromFoot = new Vector2d();
-		afterDirFromFoot.x = afterDir.y;
-		afterDirFromFoot.y = -afterDir.x;
-
-		vertex.x = footV.x + d0 * afterDirFromFoot.x;
-		vertex.y = footV.y + d0 * afterDirFromFoot.y;
-	}
-
-	private void flipFace(final OriFace face, final OriHalfedge baseHe) {
-		var baseHePair = baseHe.getPair();
-		var baseHePairNext = baseHePair.getNext();
-
-		// (Maybe) baseHe.pair keeps the position before folding.
-		Vector2d preOrigin = new Vector2d(baseHePairNext.getPositionWhileFolding());
-		Vector2d afterOrigin = new Vector2d(baseHe.getPositionWhileFolding());
-
-		// Creates the base unit vector for before the rotation
-		Vector2d baseDir = new Vector2d();
-		baseDir.sub(baseHePair.getPositionWhileFolding(), baseHePairNext.getPositionWhileFolding());
-
-		// Creates the base unit vector for after the rotation
-		var baseHeNext = baseHe.getNext();
-		Vector2d afterDir = new Vector2d();
-		afterDir.sub(baseHeNext.getPositionWhileFolding(), baseHe.getPositionWhileFolding());
-		afterDir.normalize();
-
-		Line preLine = new Line(preOrigin, baseDir);
-
-		// move the vertices of the face to keep the face connection
-		// on baseHe
-		face.halfedgeStream().forEach(he -> {
-			transformVertex(he.getPositionWhileFolding(), preLine, afterOrigin, afterDir);
-		});
-
-		face.precreaseStream().forEach(precrease -> {
-			transformVertex(precrease.p0, preLine, afterOrigin, afterDir);
-			transformVertex(precrease.p1, preLine, afterOrigin, afterDir);
-		});
-
-		// Inversion
-		if (face.isFaceFront() == baseHe.getFace().isFaceFront()) {
-			Vector2d ep = baseHeNext.getPositionWhileFolding();
-			Vector2d sp = baseHe.getPositionWhileFolding();
-
-			face.halfedgeStream().forEach(he -> {
-				flipVertex(he.getPositionWhileFolding(), sp, ep);
-			});
-			face.precreaseStream().forEach(precrease -> {
-				flipVertex(precrease.p0, sp, ep);
-				flipVertex(precrease.p1, sp, ep);
-
-			});
-			face.invertFaceFront();
-		}
-	}
-
 	/**
 	 * creates the matrix overlapRelation and fills it with "no overlap" or
 	 * "undefined"
@@ -1099,68 +987,6 @@ public class Folder {
 	 */
 	public void foldWithoutLineType(
 			final OrigamiModel model) {
-		List<OriEdge> edges = model.getEdges();
-		List<OriFace> faces = model.getFaces();
-
-//		for (OriFace face : faces) {
-//			face.faceFront = true;
-//			face.movedByFold = false;
-//		}
-
-		walkFace(faces, faces.get(0), 0);
-
-//		Collections.sort(faces, new FaceOrderComparator());
-		model.getSortedFaces().clear();
-		model.getSortedFaces().addAll(faces);
-
-		for (OriEdge e : edges) {
-			var sv = e.getStartVertex();
-			sv.getPosition().set(e.getLeft().getPositionWhileFolding());
-		}
-
-		folderTool.setFacesOutline(faces);
-	}
-
-	/**
-	 * Make the folds by flipping the faces
-	 *
-	 * @param faces
-	 * @param face
-	 * @param walkFaceCount
-	 */
-	private void walkFace(final List<OriFace> faces, final OriFace face, final int walkFaceCount) {
-		face.setMovedByFold(true);
-		if (walkFaceCount > 1000) {
-			logger.error("walkFace too deep");
-			return;
-		}
-		face.halfedgeStream().forEach(he -> {
-			var pair = he.getPair();
-			if (pair == null) {
-				return;
-			}
-			var pairFace = pair.getFace();
-			if (pairFace.isMovedByFold()) {
-				return;
-			}
-
-			flipFace2(faces, pairFace, he);
-			pairFace.setMovedByFold(true);
-			walkFace(faces, pairFace, walkFaceCount + 1);
-		});
-	}
-
-	private void flipFace2(final List<OriFace> faces, final OriFace face,
-			final OriHalfedge baseHe) {
-		flipFace(face, baseHe);
-		faces.remove(face);
-		faces.add(face);
-	}
-
-	private void flipVertex(final Vector2d vertex, final Vector2d sp, final Vector2d ep) {
-		var v = GeomUtil.getSymmetricPoint(vertex, sp, ep);
-
-		vertex.x = v.x;
-		vertex.y = v.y;
+		simpleFolder.foldWithoutLineType(model);
 	}
 }
